Update patch set 6
Patch Set 6: Code-Review-1

Why are you removing the regulatory information? Same with the csp item...

And none of these are even captured in your commit message. Really, this should be two separate commits, one for the items in the cm.mk (which is +2) and one for the items you are deleting (-2). 

Overall, as-is, -1

Patch-set: 6