Update patch set 1
Patch Set 1:

> I was under the assumption this was inherited from the non-lte
 > flounder configs.

Unfortunately it is not. My builds were using the 800 bootanimation instead of the 1536 like flounder. These values are set in cm.mk in flounder, which isn't included by flounder_lte. With this fix I get the correct animation (since I'm only targeting the Nexus 9, I only replaced the 1536 anim, so it was pretty obvious).

If there is a better way to do this, to make it inherit, that would be better. I'm just not familiar enough with everything about CM to know the best way. :)

Patch-set: 1